# backend/docs/swagger.yaml
basePath: /api
definitions:
  rest.Err:
    properties:
      code:
        type: string
      details:
        items:
          $ref: '#/definitions/rest.ErrDetail'
        type: array
      message:
        type: string
      request_id:
        type: string
    type: object
  rest.ErrDetail:
    properties:
      code:
        type: string
      field:
        type: string
      message:
        type: string
    type: object
  rest.ErrResponse:
    properties:
      error:
        $ref: '#/definitions/rest.Err'
    type: object
...
  rest.ProjectMemberListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/rest.ProjectMemberResponse'
        type: array
    type: object
  rest.ProjectMemberResponse:
    properties:
      added_by:
        type: string
      created_at:
        type: integer
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      project_id:
        type: string
      role_id:
        type: string
      role_name:
        type: string
      updated_at:
        type: integer
      user_id:
        type: string
    type: object
  rest.ProjectMemberUpsertRequest:
    properties:
      role_id:
        type: string
    type: object
...
  rest.ProjectRevisionListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/rest.ProjectRevisionResponse'
        type: array
      next_page_token:
        type: string
    type: object
  rest.ProjectRevisionResponse:
    properties:
      changed_at:
        type: integer
      changed_by:
        type: string
      created_at:
        type: integer
      deleted_at:
        type: integer
      description:
        type: string
      end_date:
        type: string
      goal:
        type: string
      id:
        type: string
      priority:
        type: string
      revision:
        type: integer
      start_date:
        type: string
      title:
        type: string
      type:
        type: string
      updated_at:
        type: integer
      user_id:
        type: string
    type: object
...
info:
  contact: {}
  description: Task scheduling API server.
  title: task2todaytodo API
  version: "1.0"
paths:
...
  /projects/{id}/members:
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.ProjectMemberListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ErrResponse'
      security:
      - BearerAuth: []
      summary: List project members
      tags:
      - Projects
  /projects/{id}/members/{user_id}:
    delete:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Member user ID
        in: path
        name: user_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ErrResponse'
      security:
      - BearerAuth: []
      summary: Remove a project member
      tags:
      - Projects
    put:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Member user ID
        in: path
        name: user_id
        required: true
        type: string
      - description: Member role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rest.ProjectMemberUpsertRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ErrResponse'
      security:
      - BearerAuth: []
      summary: Add or update a project member
      tags:
      - Projects
  /projects/{id}/revisions:
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Items per page
        in: query
        name: page_size
        type: integer
      - description: Opaque next page token
        in: query
        name: page_token
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.ProjectRevisionListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/rest.ErrResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ErrResponse'
      security:
      - BearerAuth: []
      summary: List project revisions
      tags:
      - Projects
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"